test(governance): 拆分线程历史加载测试
